+++ softwareconfiguration/autofe/src/autofe/algorithm/hasco/filter/meta/FilterPipelineFactory.java
@@ -20,9 +20,9 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 
 	private static final Logger logger = LoggerFactory.getLogger(FilterPipelineFactory.class);
 
-	private final int[] inputShape;
+	private final long[] inputShape;
 
-	public FilterPipelineFactory(final int[] inputShape) {
+	public FilterPipelineFactory(final long[] inputShape) {
 		this.inputShape = inputShape;
 	}
 
@@ -45,7 +45,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 				return new FilterPipeline(null);
 			}
 
-			IFilter actCIFilter = FilterUtils.getFilterForName(actCI.getComponent().getName(), actCI.getParameterValues(), this.inputShape);
+			IFilter actCIFilter = FilterUtils.getFilterForName(actCI.getComponent().getName(),
+					actCI.getParameterValues(), this.inputShape);
 			filterGraph.addItem(actCIFilter);
 			openFilter.offer(actCIFilter);
 			open.offer(actCI);
@@ -66,7 +67,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 				case UNION_NAME:
 					ComponentInstance filter1CI = actCI.getSatisfactionOfRequiredInterfaces().get("filter1");
 					if (filter1CI != null) {
-						IFilter filter1 = FilterUtils.getFilterForName(filter1CI.getComponent().getName(), filter1CI.getParameterValues(), this.inputShape);
+						IFilter filter1 = FilterUtils.getFilterForName(filter1CI.getComponent().getName(),
+								filter1CI.getParameterValues(), this.inputShape);
 
 						open.offer(filter1CI);
 						openFilter.offer(filter1);
@@ -78,7 +80,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 
 					ComponentInstance filter2CI = actCI.getSatisfactionOfRequiredInterfaces().get("filter2");
 					if (filter2CI != null) {
-						IFilter filter2 = FilterUtils.getFilterForName(filter2CI.getComponent().getName(), filter2CI.getParameterValues(), this.inputShape);
+						IFilter filter2 = FilterUtils.getFilterForName(filter2CI.getComponent().getName(),
+								filter2CI.getParameterValues(), this.inputShape);
 
 						open.offer(filter2CI);
 						openFilter.offer(filter2);
@@ -100,7 +103,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 							break;
 						}
 
-						extractor = FilterUtils.getFilterForName(extractorCI.getComponent().getName(), extractorCI.getParameterValues(), this.inputShape);
+						extractor = FilterUtils.getFilterForName(extractorCI.getComponent().getName(),
+								extractorCI.getParameterValues(), this.inputShape);
 						filterGraph.addItem(extractor);
 						filterGraph.addEdge(actCIFilter, extractor);
 
@@ -111,7 +115,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 							break;
 						}
 
-						extractor = FilterUtils.getFilterForName(extractorCI.getComponent().getName(), extractorCI.getParameterValues(), this.inputShape);
+						extractor = FilterUtils.getFilterForName(extractorCI.getComponent().getName(),
+								extractorCI.getParameterValues(), this.inputShape);
 
 						// If pretrained neural net can not be applied to given input shape (e. g. due
 						// to different channel amount) return empty filter pipeline (checked in node
@@ -136,20 +141,23 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 					IFilter preprocessor = null;
 					if (!preprocessorCI.getComponent().getName().equals("PrepPipe")) {
 						// Just one basic filter
-						preprocessor = FilterUtils.getFilterForName(preprocessorCI.getComponent().getName(), preprocessorCI.getParameterValues(), this.inputShape);
+						preprocessor = FilterUtils.getFilterForName(preprocessorCI.getComponent().getName(),
+								preprocessorCI.getParameterValues(), this.inputShape);
 						filterGraph.addItem(preprocessor);
 						filterGraph.addEdge(actCIFilter, preprocessor);
 					} else {
 						// Preprocessor pipeline
 						IFilter newActCIFilter = actCIFilter;
 						while (preprocessorCI != null && preprocessorCI.getComponent().getName().equals("PrepPipe")) {
-							ComponentInstance childCI = preprocessorCI.getSatisfactionOfRequiredInterfaces().get("preprocessor");
+							ComponentInstance childCI = preprocessorCI.getSatisfactionOfRequiredInterfaces()
+									.get("preprocessor");
 
 							if (childCI == null) {
 								break;
 							}
 
-							IFilter childFilter = FilterUtils.getFilterForName(childCI.getComponent().getName(), childCI.getParameterValues(), this.inputShape);
+							IFilter childFilter = FilterUtils.getFilterForName(childCI.getComponent().getName(),
+									childCI.getParameterValues(), this.inputShape);
 							filterGraph.addItem(childFilter);
 							filterGraph.addEdge(newActCIFilter, childFilter);
 
@@ -159,7 +167,8 @@ public class FilterPipelineFactory implements Factory<FilterPipeline> {
 
 						// End of pipeline reached
 						if (preprocessorCI != null) {
-							preprocessor = FilterUtils.getFilterForName(preprocessorCI.getComponent().getName(), preprocessorCI.getParameterValues(), this.inputShape);
+							preprocessor = FilterUtils.getFilterForName(preprocessorCI.getComponent().getName(),
+									preprocessorCI.getParameterValues(), this.inputShape);
 							filterGraph.addItem(preprocessor);
 							filterGraph.addEdge(newActCIFilter, preprocessor);
 						}
