+++ softwareconfiguration/autofe/src/autofe/util/DataSetUtils.java
@@ -41,13 +41,13 @@ public final class DataSetUtils {
 	public static final String API_KEY = "4350e421cdc16404033ef1812ea38c01";
 
 	public static final int CIFAR10_ID = 40927;
-	public static final int[] CIFAR10_INPUT_SHAPE = new int[] { 32, 32, 3 };
+	public static final long[] CIFAR10_INPUT_SHAPE = new long[] { 32, 32, 3 };
 	public static final int SEGMENT_ID = 40984;
 	public static final int MNIST_ID = 554;
-	public static final int[] MNIST_INPUT_SHAPE = new int[] { 28, 28, 1 };
+	public static final long[] MNIST_INPUT_SHAPE = new long[] { 28, 28, 1 };
 	public static final int FASHION_MNIST_ID = 40996;
 	// public static final int[] FASHION_MNIST_SHAPE = new int[] { 1, 1, 28, 28 };
-	public static final int[] FASHION_MNIST_SHAPE = new int[] { 28, 28, 1 };
+	public static final long[] FASHION_MNIST_SHAPE = new long[] { 28, 28, 1 };
 
 	private static final Logger logger = LoggerFactory.getLogger(DataSetUtils.class);
 
@@ -238,7 +238,7 @@ public final class DataSetUtils {
 		return inst;
 	}
 
-	public static int[] getInputShapeByDataSet(final int dataSetID) {
+	public static long[] getInputShapeByDataSet(final int dataSetID) {
 		switch (dataSetID) {
 		case MNIST_ID:
 			return MNIST_INPUT_SHAPE;
@@ -247,7 +247,7 @@ public final class DataSetUtils {
 		case FASHION_MNIST_ID:
 			return MNIST_INPUT_SHAPE;
 		default:
-			return new int[] { 1 };
+			return new long[] { 1 };
 		}
 	}
