+++ softwareconfiguration/autofe/test/autofe/test/ImageAutoFETest.java
@@ -40,7 +40,7 @@ public class ImageAutoFETest {
 	private static final int MAX_PIPELINE_SIZE = 5;
 
 	private static final int USED_DATASET = DataSetUtils.MNIST_ID;
-	private static final int[] DATASET_INPUT_SHAPE = DataSetUtils.getInputShapeByDataSet(USED_DATASET);
+	private static final long[] DATASET_INPUT_SHAPE = DataSetUtils.getInputShapeByDataSet(USED_DATASET);
 
 	private static final boolean ENABLE_MLPLAN_VIS = true;
 
@@ -63,8 +63,9 @@ public class ImageAutoFETest {
 		}
 		logger.info("Finished intermediate calculations.");
 
-		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"), new ClusterNodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate), new ClusterObjectEvaluator(),
-				DATASET_INPUT_SHAPE);
+		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"),
+				new ClusterNodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate),
+				new ClusterObjectEvaluator(), DATASET_INPUT_SHAPE);
 		// hascoFE.enableVisualization();
 		hascoFE.setLoggerName("autofe image");
 		hascoFE.runSearch(AUTOFE_TIMEOUT);
@@ -84,11 +85,15 @@ public class ImageAutoFETest {
 				intermediateSplit1.add(DataSetUtils.instanceToMatrixByDataSet(inst, USED_DATASET));
 			}
 
-			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0), false);
-			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1), false);
+			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0),
+					false);
+			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1),
+					false);
 
-			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(), resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
-			System.out.println("Error Rate of the solution produced by ML-Plan (Cluster): " + (100 - mlPlanResult) / 100f);
+			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(),
+					resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
+			System.out.println(
+					"Error Rate of the solution produced by ML-Plan (Cluster): " + (100 - mlPlanResult) / 100f);
 
 		} else {
 			logger.info("No solution could be found.");
@@ -115,8 +120,9 @@ public class ImageAutoFETest {
 		}
 		logger.info("Finished intermediate calculations.");
 
-		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"), new LDANodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate), new LDAObjectEvaluator(),
-				DATASET_INPUT_SHAPE);
+		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"),
+				new LDANodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate),
+				new LDAObjectEvaluator(), DATASET_INPUT_SHAPE);
 		// hascoFE.enableVisualization();
 		hascoFE.setLoggerName("autofe image");
 		hascoFE.runSearch(AUTOFE_TIMEOUT);
@@ -136,12 +142,15 @@ public class ImageAutoFETest {
 				intermediateSplit1.add(DataSetUtils.instanceToMatrixByDataSet(inst, USED_DATASET));
 			}
 
-			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0), false);
-			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1), false);
+			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0),
+					false);
+			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1),
+					false);
 
 			logger.info("Feature size of generated solution: " + resultSplit0.getInstances().numAttributes());
 
-			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(), resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
+			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(),
+					resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
 			System.out.println("Error Rate of the solution produced by ML-Plan (LDA): " + (100 - mlPlanResult) / 100f);
 
 		} else {
@@ -156,7 +165,8 @@ public class ImageAutoFETest {
 
 		long timeStart = System.currentTimeMillis();
 
-		List<Instances> dataSetVariations = HASCOFeatureEngineering.generateRandomDataSets(USED_DATASET, 1, MAX_PIPELINE_SIZE, AUTOFE_TIMEOUT, 42);
+		List<Instances> dataSetVariations = HASCOFeatureEngineering.generateRandomDataSets(USED_DATASET, 1,
+				MAX_PIPELINE_SIZE, AUTOFE_TIMEOUT, 42);
 		// OpenmlConnector connector = new OpenmlConnector();
 		// DataSetDescription ds = connector.dataGet(USED_DATASET);
 		// File file = ds.getDataset(DataSetUtils.API_KEY);
@@ -168,7 +178,8 @@ public class ImageAutoFETest {
 		long timeDataSetVar = System.currentTimeMillis();
 		System.out.println("timeDataSetVar: " + timeDataSetVar);
 
-		double mlPlanScore = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, dataSetVariations.get(0), 0.75, 42, logger, false, 6);
+		double mlPlanScore = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, dataSetVariations.get(0), 0.75, 42, logger,
+				false, 6);
 		System.out.println("MLPLan Score: " + mlPlanScore);
 
 		long endTime = System.currentTimeMillis();
@@ -197,8 +208,9 @@ public class ImageAutoFETest {
 		}
 		logger.info("Finished intermediate calculations.");
 
-		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"), new EnsembleNodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate), new EnsembleObjectEvaluator(),
-				DATASET_INPUT_SHAPE);
+		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"),
+				new EnsembleNodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate),
+				new EnsembleObjectEvaluator(), DATASET_INPUT_SHAPE);
 		// hascoFE.enableVisualization();
 		hascoFE.setLoggerName("autofe image");
 		hascoFE.runSearch(AUTOFE_TIMEOUT);
@@ -218,13 +230,17 @@ public class ImageAutoFETest {
 				intermediateSplit1.add(DataSetUtils.instanceToMatrixByDataSet(inst, USED_DATASET));
 			}
 
-			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0), false);
-			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1), false);
+			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0),
+					false);
+			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1),
+					false);
 
 			logger.info("Feature size of generated solution: " + resultSplit0.getInstances().numAttributes());
 
-			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(), resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
-			System.out.println("Error Rate of the solution produced by ML-Plan (Ensemble): " + (100 - mlPlanResult) / 100f);
+			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(),
+					resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
+			System.out.println(
+					"Error Rate of the solution produced by ML-Plan (Ensemble): " + (100 - mlPlanResult) / 100f);
 
 		} else {
 			logger.info("No solution could be found.");
@@ -250,8 +266,9 @@ public class ImageAutoFETest {
 		}
 		logger.info("Finished intermediate calculations.");
 
-		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"), new COCONodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate), new COCOObjectEvaluator(),
-				DATASET_INPUT_SHAPE);
+		HASCOFeatureEngineering hascoFE = new HASCOFeatureEngineering(new File("model/catalano/catalano.json"),
+				new COCONodeEvaluator(MAX_PIPELINE_SIZE), new DataSet(split.get(0), intermediate),
+				new COCOObjectEvaluator(), DATASET_INPUT_SHAPE);
 		// hascoFE.enableVisualization();
 		hascoFE.setLoggerName("autofe image");
 		hascoFE.runSearch(AUTOFE_TIMEOUT);
@@ -271,12 +288,15 @@ public class ImageAutoFETest {
 				intermediateSplit1.add(DataSetUtils.instanceToMatrixByDataSet(inst, USED_DATASET));
 			}
 
-			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0), false);
-			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1), false);
+			DataSet resultSplit0 = solution.getSolution().applyFilter(new DataSet(newSplit.get(0), intermediateSplit0),
+					false);
+			DataSet resultSplit1 = solution.getSolution().applyFilter(new DataSet(newSplit.get(1), intermediateSplit1),
+					false);
 
 			logger.info("Feature size of generated solution: " + resultSplit0.getInstances().numAttributes());
 
-			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(), resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
+			double mlPlanResult = EvaluationUtils.evaluateMLPlan(MLPLAN_TIMEOUT, resultSplit0.getInstances(),
+					resultSplit1.getInstances(), 42, logger, ENABLE_MLPLAN_VIS);
 			System.out.println("Error Rate of the solution produced by ML-Plan (COCO): " + (100 - mlPlanResult) / 100f);
 
 		} else {
